AIWearPython/requirements.txt
dashscope==1.25.9
flask==3.1.3
flask-cors==4.0.0
langchain-community==0.4.1
langchain-core==1.2.7
pillow==12.1.0
python-dotenv==1.2.1